page_name AND fn to set per-page metadata
Enrich rendering context by providing the page's path - relative
to the pages directory.
Furthermore, provide function to update a dict containing per-page
metadata.
This permits dynamically generated pages (and the like) to use this
information to generate pages whose contents depend on data provided
by the regular page templates.
The primary motivation for this feature will be for pages to provide
a list of tags and for these to be collected in per-page meta-data
entries which tag index pages can then use to render the page.