rewrite - use multiprocessing
Use multiprocessing to spawn a compile process which itself
receives events ({PAGE,TEMPLATE,CONTEXT}_CHANGED) triggering
selective or full recompilation of all pages.

To handle a change in any of the context code files, the process
is terminated and restarted by its handler. This way, the context
modules are not in cache and so the new code will be executed.

Otherwise, the compile process (and context!) is reused between
events to improve performance.